+++ include/com.h
@@ -1,0 +1,11 @@
+#pragma once
+
+#include <com/helpers.h>
+#include <com/ComPtr.h>
+#include <com/ComClass.h>
+#include <com/IDispatchImpl.h>
+#include <com/IOleInPlaceActiveObjectImpl.h>
+#include <com/IOleInPlaceObjectImpl.h>
+#include <com/IOleObjectImpl.h>
+#include <com/IUnknownImpl.h>
+#include <com/IViewObjectExImpl.h>
