+++ shared/Wow64.h
@@ -2,20 +2,12 @@
 
 #include <windows.h>
 
-#ifdef __cplusplus
-extern "C" {
-#endif
-
 #if _WIN64
 	#define OurGetNativeSystemInfo(lpSystemInfo) GetNativeSystemInfo(lpSystemInfo)
 	#define DisableWow64FsRedirection(OldValue)  TRUE
 	#define RevertWow64FsRedirection(OldValue)   TRUE
 #else
-	void OurGetNativeSystemInfo(LPSYSTEM_INFO lpSystemInfo);
-	BOOL DisableWow64FsRedirection(PVOID *OldValue);
-	BOOL RevertWow64FsRedirection(PVOID OldValue);
-#endif
-
-#ifdef __cplusplus
-}
+	EXTERN_C void OurGetNativeSystemInfo(LPSYSTEM_INFO lpSystemInfo);
+	EXTERN_C BOOL DisableWow64FsRedirection(PVOID *OldValue);
+	EXTERN_C BOOL RevertWow64FsRedirection(PVOID OldValue);
 #endif
